-Pequenas alterações

+++ src/java/bolsav/StockCli.java
@@ -14,7 +14,7 @@ public class StockCli implements Serializable {
 
     public Stock stock;
     public String client;
-    public long id;
+    public int id;
     //variável para ver se a compra realizou ou não
     public boolean statusSell;
 
@@ -26,7 +26,7 @@ public class StockCli implements Serializable {
      * @param client com a referência do cliente
      * @param id com o id do cliente
      */
-    public StockCli(Stock stock, String client, long id) {
+    public StockCli(Stock stock, String client, int id) {
         this.stock = stock;
         this.client = client;
         this.id = id;
@@ -58,7 +58,7 @@ public class StockCli implements Serializable {
      *
      * @return do tipo long com o id do cliente
      */
-    public long getId() {
+    public int getId() {
         return id;
     }
